Move count utils to tracking module.
